Chore: initializing log model on main file and fix on all imports
