+++ src/models/entity.model.ts
@@ -1,17 +1,6 @@
 import { DataTypes } from 'sequelize'
 import { storageSequelize } from '../database/connection'
 
-export type Entity = {
-    id: number
-    title: string
-    properties: object
-    description: string
-    author?: string
-    image?: object
-    sections?: string
-    type: string
-}
-
 export const EntityModel = storageSequelize.define('entity', {
     id: {
         type: DataTypes.INTEGER,
