+++ src/models/types/Entity.ts
@@ -1,0 +1,10 @@
+export type Entity = {
+    id: number
+    title: string
+    properties: object
+    description: string
+    author?: string
+    image?: object
+    sections?: string
+    type: string
+}
