+++ src/models/types/Log.ts
@@ -1,0 +1,4 @@
+export type Log = {
+    type: string
+    content: string
+}
