+++ src/models/types/User.ts
@@ -1,0 +1,6 @@
+export type User = {
+    id: number
+    name: string
+    username: string
+    password: string
+}
